hyperagent/requirements.txt
textual>=8.2.0
textual-plotext
hyperliquid-python-sdk>=0.23.0
anthropic>=0.28.0
boto3
numpy
ta
python-dotenv
eth-account
